requirements.txt
Flask==3.1.0
Flask-Babel==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail>=0.9.0
Flask-WTF==1.2.1
email-validator==2.1.0.post1
python-dotenv==1.0.1
PyJWT>=2.0.0
bcrypt>=3.2.0
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
SQLAlchemy==2.0.27
